mv fixtures to conftest.py file
